requirement.txt
aniso8601==3.0.2
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2018.4.16
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-datatables==0.4.0
djangorestframework-jsonapi==2.5.0
djangorestframework-recursive==0.1.2
feedparser==5.1.3
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
inflection==0.3.1
iso8601==0.1.12
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
oauthlib==1.0.3
obd==0.6.1
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
Pint==0.8.1
pipenv==2018.7.1
plainbox==0.25
promise==2.1
ptyprocess==0.5
pyasn1==0.1.9
PyBluez==0.22
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
pyserial==3.4
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-systemd==231
pytz==2018.5
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
Rx==1.6.1
sessioninstaller==0.0.0
singledispatch==3.4.0.3
six==1.10.0
system-service==0.3
typing==3.6.4
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.2.0
virtualenv-clone==0.3.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
